moved classes out of products class into individual classes
